fix issues

+++ 0x13-more_singly_linked_lists/0-print_listint.c
@@ -11,7 +11,7 @@
 size_t print_listint(const listint_t *h)
 {
 	listint_t *new;
-	unsigned int count = 0;
+	size_t count = 0;
 
 	new = h;
 	while (new != NULL)
